sc_options: add explicit collective mode
Historically, we expect the sc_options functions to be called collectively.
This is reflected in using the SC_LC_GLOBAL log category for usage and summary.
It was not reflected in the file load/save functions which relied on the caller.

Now we add a set_collective function to enable/disable collective logging and I/O.