Add missing C linkage to sc_extern_c_hack_2
We declare dummy prototypes to cater to the needs of GNU indent
when it encounters our SC_EXTERN_C_BEGIN macros plus a semicolon,
and likewise for SC_EXTERN_C_END.  To make sure the semicolon is
not semantically superfluous, we add said prototypes.

These are now made the same for both C and C++ compile of libsc.