feature-scda: Introduce two versions of mod padding
Since we do not want to copy the potentially very big data, we also
offer a padding function that only sets the padding bytes but does
not copy the input data. For the fixed-length padding the data sizes are
smaller and known in advance. That is why, we prefer the inplace
padding for the fixed-length padding. However, for the sakes of
flexibility and symmetry we will also add non-inplace function for
the fixed-length padding.